Updated NY training set and code
Training set now excludes records matched as true but with more than 5 year difference in age (this is due to being incorrectly linked to a father / something like that). Also is attached is some quick code I threw together to run the training set through a Random Forest. Can't get XGBoost to install on the school server, so I couldn't do that, but Random Forest is good enough. I get an accuracy of ~97.75. Almost all of the inaccuracy is driven by false negatives (i.e. two given records are indeed matches but the model says that they aren't matches). This is could be due to surname changes by females or abbreviations / nicknames (e.g. Wm. and William are the same name but score vastly different with bigram scoring).

The next step is to get a training set for a state that is pretty different (e.g. Utah) and  see how the model for one predicts the matches in another. 

The next code I upload will be a lot cleaner and compare Random Forest, XGBoost, Neural Networks, and Sup